KObject: Object Identities
Object Identities are unique objects that are compared when cloning an
existing object back to another process. Before this, ReadStreams, when
duplicated then passed to a process with a ReadStream with the same
data, instead of giving out a new ReadStream, it would use the old one.
This was apparent when calling term.read multiple times: Even if you
duplicated the stream (in my case, I did so twice, once in process Zero,
and again in KAPI), you would always get the same object back (which
could be closed).