Zero: Implement Permissions Model
Filesystem permissions are currently not implemented on the filesystem
side, but the rest of the system is.

Permissions have targets. These targets are defined by strings in a
special format. perm.query returns which upper level target accepts or
denies a permission.

Permissions themselves are hierarchal. test.a.perm will first check
"test.a.perm", then "test.a", then finally "test" to figure out if a
permission has been allowed or denied.

Targets are also hierarchal. A process target will query as a user
target if undefined. A user target (which can also be a process target)
will query as a group target if undefined. If none of the targets
specify a permission, the result is returned nil, and the location is
"undefined".

If multiple groups have conflicting permissions, the result is
undefined.

perm.list is not defined yet.