KFuture: Isolate errors in various situations where objects are deleted
KFutures exhibited a denial of service type attack. If you deleted a
future object from another process before it gets completed, the
process, when trying to complete the future, will crash. This caused
crashes down the chain, until the only process left was terminal, with a
blinking cursor.

This fixes it by checking the validity of objects before going through
with operations. A delete method has also been added to Future, allowing
it to disconnect the end point.

If the completer attached to a future callback gets deleted (by the
process being killed), a denial of service error happens.

If the future is deleted while inside of it's callback (the process
being killed), it will cause a denial of service error.